Add models submodule
