+++ duqtools/ids/_location.py
@@ -9,7 +9,8 @@ from typing import TYPE_CHECKING
 import imas
 from imas import imasdef
 
-from .._types import BaseModel
+from duqtools.models import BaseModel
+
 from ._mapping import IDSMapping
 
 if TYPE_CHECKING:
