+++ duqtools/models/__init__.py
@@ -1,0 +1,7 @@
+from ._basemodel import BaseModel
+from ._runs import Runs
+
+__all__ = [
+    'Runs',
+    'BaseModel',
+]
