+++ duqtools/models/_basemodel.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel as PydanticBaseModel
+from pydantic import Extra
+
+
+class BaseModel(PydanticBaseModel):
+
+    class Config:
+        extra = Extra.forbid
